requirements.txt
Django==5.1.5
djangorestframework==3.17.1
django-cors-headers==4.9.0
mysqlclient==2.2.8
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.2.2
drf-spectacular==0.29.0
djangorestframework_simplejwt==5.5.1
celery==5.6.3
redis==7.4.0
qrcode==8.2
reportlab==4.4.4
django-auditlog==2.3.0
firebase-admin==6.5.0
